parse ld efuns
