+++ .vscode/launch.json
@@ -12,8 +12,8 @@
                 //"${workspaceRoot}/../lima",
                 //"${workspaceRoot}/../StickLib",
                 //"${workspaceRoot}/../nightmare-residuum",
-                //"${workspaceRoot}/../lp-245",
-                "${workspaceRoot}/efuns/ldmud",
+                "${workspaceRoot}/../lp-245",
+                //"${workspaceRoot}/efuns/ldmud",
                 "--extensionDevelopmentPath=${workspaceRoot}"
             ],
             "autoAttachChildProcesses": true,
